api/requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
python-dotenv==0.20.0
python-multipart==0.0.5
SQLAlchemy==1.4.37
pydantic==1.10.2
alembic==1.8.0
python-jose~=3.3.0
psycopg2-binary==2.9.3
passLib==1.7.4
fastapi-mail==1.1.4
schedule==1.1.0
python_dateutil==2.8.2

pytest==7.1.2
pytest-cov==3.0.0
requests==2.28.0
pre-commit==2.19.0
flake8==5.0.4
black==22.6.0
isort==5.10.1
